+++ release/githubio/js/main_uncompressed.js
@@ -1775,9 +1775,15 @@
         VersionBase: '3.0',
         publishVersionCode: 310, //Always update (+1 to current value) Main_version_java after update publishVersionCode or a major update of the apk is released
         ApkUrl: 'https://github.com/fgl27/SmartTwitchTV/releases/download/310/SmartTV_twitch_3_0_310.apk',
-        WebVersion: 'February 28 2020',
-        WebTag: 579, //Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
+        WebVersion: 'March 8 2020',
+        WebTag: 580, //Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
         changelog: [{
+                title: "Web Version March 8 2020",
+                changes: [
+                    "Add support to see chat messages that were sent before you joined the stream"
+                ]
+            },
+            {
                 title: "Web Version February 28 2020",
                 changes: [
                     "General improves and bug fixes"
@@ -4743,6 +4749,7 @@
         if (ChatLive_selectedChannel[chat_number]) ChatLive_selectedChannel[chat_number] = ChatLive_selectedChannel[chat_number].toLowerCase();
 
         ChatLive_SetOptions(chat_number, Chat_Id[chat_number]);
+        ChatLive_PreLoadChat(chat_number, Chat_Id[chat_number]);
 
         ChatLive_loadChatters(chat_number, Chat_Id[chat_number]);
         ChatLive_loadEmotesUser();
@@ -5305,6 +5312,43 @@
         }
     }
 
+    function ChatLive_PreLoadChat(chat_number, id) {
+
+        BaseXmlHttpGet(
+            'https://recent-messages.robotty.de/api/v2/recent-messages/' + ChatLive_selectedChannel[chat_number] + '?limit=30',
+            0,
+            null,
+            ChatLive_PreLoadChatSuccess,
+            noop_fun,
+            chat_number,
+            id
+        );
+
+    }
+
+    function ChatLive_PreLoadChatSuccess(data, chat_number, id) {
+
+        if (id !== Chat_Id[chat_number]) return;
+
+        var obj = JSON.parse(data),
+            len = obj.messages.length,
+            i = 0,
+            message;
+
+        for (i; i < len; i++) {
+
+            message = window.parseIRC(obj.messages[i].trim());
+
+            if (message.command === "PRIVMSG") {
+
+                ChatLive_loadChatSuccess(message, chat_number, true);
+
+            }
+
+        }
+
+    }
+
     var useToken = [];
 
     function ChatLive_loadChat(chat_number, id) {
@@ -5314,8 +5358,7 @@
 
         ChatLive_LineAdd({
             chat_number: chat_number,
-            message: ChatLive_LineAddSimple(STR_LOADING_CHAT + STR_SPACE + STR_LIVE + STR_SPACE + STR_CHANNEL + ': ' +
-                (!chat_number ? Play_data.data[1] : PlayExtra_data.data[1]))
+            message: ChatLive_LineAddSimple(STR_LOADING_CHAT + STR_SPACE + (!chat_number ? Play_data.data[1] : PlayExtra_data.data[1])) + STR_SPACE + STR_LIVE
         });
 
         useToken[chat_number] = ChatLive_Logging && !ChatLive_Banned[chat_number] && AddUser_IsUserSet() && AddUser_UsernameArray[0].access_token;
@@ -5972,7 +6015,7 @@
         );
     }
 
-    function ChatLive_loadChatSuccess(message, chat_number) {
+    function ChatLive_loadChatSuccess(message, chat_number, addToStart) {
         var div = '',
             tags = message.tags,
             nick,
@@ -6081,7 +6124,8 @@
             atstreamer: atstreamer,
             atuser: atuser,
             hasbits: (hasbits && ChatLive_Highlight_Bits),
-            extraMessage: extraMessage
+            extraMessage: extraMessage,
+            addToStart: addToStart
         };
 
         ChatLive_LineAddCheckDelay(chat_number, messageObj);
@@ -6279,7 +6323,15 @@
         // </span>
         // </div>
 
-        Chat_div[messageObj.chat_number].appendChild(elem);
+        if (!messageObj.addToStart) {
+
+            Chat_div[messageObj.chat_number].appendChild(elem);
+
+        } else {
+
+            Chat_div[messageObj.chat_number].insertBefore(elem, Chat_div[messageObj.chat_number].childNodes[0]);
+
+        }
     }
 
     function ChatLive_MessagesRunAfterPause() {
@@ -6712,8 +6764,7 @@
             Chat_MessageVector({
                 chat_number: 0,
                 time: 0,
-                message: '<span class="message">' + STR_LOADING_CHAT + Chat_title + STR_SPACE + STR_CHANNEL + ': ' +
-                    Main_values.Main_selectedChannelDisplayname + '</span>'
+                message: '<span class="message">' + STR_LOADING_CHAT + STR_SPACE + Main_values.Main_selectedChannelDisplayname + STR_SPACE + Chat_title + '</span>'
             });